services/ingestion/requirements.txt
# Buzzy Today — Ingestion Service Dependencies
feedparser==6.0.11
httpx==0.28.1
google-cloud-firestore==2.20.1
firebase-admin==6.6.0
openai==1.82.0
structlog==25.1.0
python-dotenv==1.1.0
Pillow==11.1.0
beautifulsoup4==4.13.3
lxml==5.3.1
pytest==8.3.5
pytest-asyncio==0.25.3
